% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_futility_rule.R
\name{adjust_for_futility_rule}
\alias{adjust_for_futility_rule}
\title{Adjust Dose-Response Table when Futility Rule Invoked}
\usage{
dose_info = adjust_for_futility_rule(60, dose_info) #if dose level 60 invoked futility rule
}
\arguments{
\item{dose}{A numeric dose level that violated the futility rule}

\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}
}
\value{
Updated dose-response table dataframe with adjusted availabilities
}
\description{
Adjust dose availabilities given that the futility rule was invoked, if dose sample size >= 6 (if sample size < 6, choose to stay and explore dose further, so no availabilities are changed).
The dose that invoked the futility rule and any higher doses will be made unavailable for future consideration.
}
