% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_safety_rule.R
\name{adjust_for_safety_rule}
\alias{adjust_for_safety_rule}
\title{Adjust Dose-Response Table when Safety Rule Invoked}
\usage{
dose_info = adjust_for_safety_rule(120, dose_info) #if dose level 120 invoked safety rule
}
\arguments{
\item{dose}{A numeric dose level that violated the safety rule}

\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}
}
\value{
Updated dose-response table dataframe with adjusted availabilities
}
\description{
Adjust dose availabilities given that the safety rule was invoked, if dose sample size >= 6 (if sample size < 6, choose to stay and explore dose further, so no availabilities are changed).
The dose that invoked the safety rule and any higher doses will be made unavailable for future consideration.
}
