% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_utility.R
\name{calc_utility}
\alias{calc_utility}
\title{Calculate Utility}
\usage{
calc_utility(dose_info$X_d[1], dose_info$Y_d[1], dose_info$Z_d[1], num_samples = 1e4)
}
\arguments{
\item{a}{Futility count}

\item{b}{Efficacy count}

\item{c}{Toxicity count}

\item{num_samples}{Number of times to sample Dirichlet(a,b,c) distribution. The higher this number, the more precise the estimate (default = 1e4).}
}
\value{
The calculated utility value (numeric)
}
\description{
Calculate expected utility function = 1 - E[lower loss] - E[upper loss]
}
\details{
The observed pf and pt values are only one instantiation of the Dirichlet(a,b,c) distribution.
In this function we simulate many draws from this distribution and calculate expected losses by averaging loss over all drawn toxicity and futility probabilities, and then derive the Utility value for that dose level.
In other words, this calculation is mathematically equivalent to finding the lower and upper utility values for every sample,
and then taking Utility = 1 - mean(all toxicity losses) - mean(all futility losses).

Note that the bottleneck of this function is \code{num_samples} -- the higher this value is, the slower the overall simulation performance. Between 1e4 and 1e6 is a reasonable range.
}
