% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_decisiontable_entry.R
\name{change_decisiontable_entry}
\alias{change_decisiontable_entry}
\title{Smooth Decision Table}
\usage{
decision_table15 = make_decision_table(15, PF = 0.2, PF_tolerance = 0.05, eta = 0.8, PT = 0.2, PT_tolerance = 0.05, zeta = 0.8)
decision_table15 = change_decisiontable_entry(decision_table15, 12, 4, "E")
}
\arguments{
\item{decision_table}{A matrix decision table, returned from \code{make_decision_table()}}

\item{futility_index}{Futility interval index(es) of decision to be changed}

\item{toxicity_index}{Toxicity interval index(es) of decision to be changed}

\item{new_decision}{Decision to be changed to, one of "DU", "D", "S", "E", or "EU"}
}
\value{
A matrix containing the updated decision table
}
\description{
Change any entry of a decision table
}
\details{
Note that the entries for futility and toxicity indexes can be single values or vectors. However, when changing entries values, this will change all the entries within the bounds of the specified ranges which can lead to unexpected function behavior.
For usage safety, please only enter one entry to change at a time. Always make sure to check the smoothed decision table afterwards to make sure the results are as desired.
}
