% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_dose_info.R
\name{change_dose_info}
\alias{change_dose_info}
\title{Change Dose-Response Table Row Entries}
\usage{
dose_info = change_dose_info(80, 1, 2, 0) #update dose 80 to 1 futility and 2 efficacy counts
}
\arguments{
\item{dose}{A numeric dose level (assumed to be available)}

\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}

\item{x_d}{Number of subjects experiencing futility}

\item{y_d}{Number of subjects experiencing efficacy}

\item{z_d}{Number of subjects experiencing toxicity}
}
\value{
A dataframe containing updated dose-response table (\code{dose_info})
}
\description{
Change an entire row in dose-response table with desired information.
For the given dose level, change the existing futility, efficacy, and toxicity counts to the specified numbers.
Useful for fixing entry mistakes and debugging.
}
