% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_futility_rule.R
\name{check_futility_rule}
\alias{check_futility_rule}
\title{Check Futility Rule}
\usage{
check_futility_rule(80, dose_info, 0.8)
}
\arguments{
\item{dose}{A numeric dose level (assumed to be available)}

\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}

\item{eta}{Invoke stopping rule if more than this proportion of subjects exhibit futility}
}
\value{
Boolean T/F whether futility rule was violated
}
\description{
Check if an available dose violates futility rule (futility count/sample size > futility threshold (eta).
}
