% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_safety_rule.R
\name{check_safety_rule}
\alias{check_safety_rule}
\title{Check Safety Rule}
\usage{
check_safety_rule(80, dose_info, 0.8)
}
\arguments{
\item{dose}{A numeric dose level (assumed to be available)}

\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}

\item{zeta}{Invoke stopping rule if more than this proportion of subjects exhibit toxicity}
}
\value{
Boolean T/F whether safety rule was violated
}
\description{
Check if an available dose violates safety rule (toxicity count/sample size > safety threshold zeta).
}
