% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stopping_rules.R
\name{check_stopping_rules}
\alias{check_stopping_rules}
\title{Check Stopping Rules}
\usage{
check_stopping_rules(80, dose_info, 15, 24, "D", closest_lower_dose = 60)
}
\arguments{
\item{dose}{A numeric dose level (assumed to be available)}

\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}

\item{sample_size}{Current sample size (across all doses)}

\item{max_samp_size}{Maximum sample size for entire trial}

\item{decision}{Decision, one of "DU", "D", "S", "E", or "EU". Defaults to \code{NA}.}

\item{closest_lower_dose}{If decision is to de-escalate, the closest available lower dose, \code{-Inf} if none available. Defaults to \code{NA}.}

\item{closest_higher_dose}{If decision is to escalate, the closest available higher dose, \code{Inf} if none available. Defaults to \code{NA}.}
}
\value{
Boolean T/F whether any stopping rules were violated
}
\description{
check if dose invokes stopping rules:
\itemize{
\item Maximum sample size reached
\item No doses available for evaluation
\item Lowest dose is too toxic
\item Highest dose is too futile
\item Decision is DU/EU and no doses in direction to move are available for evaluation
}
}
