% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_decision_table.R
\name{find_decision_table}
\alias{find_decision_table}
\title{Find Decision Table}
\usage{
decision_table = find_decision_table(15) #returns `decision_table15` into `decision_table` variable
}
\arguments{
\item{dose_samp_size}{Sample size to extract decision table for}
}
\value{
A matrix containing the correct decision table
}
\description{
Given the dose sample size, choose the right (pre-calculated) decision table to get decision from.
Decision tables are created using \code{make_decision_table()} and must be named "decision_table#" where #=N.
Variable is read from current environment, so decision table must be loaded.
}
