% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_decision.R
\name{make_decision}
\alias{make_decision}
\title{Make Decision}
\usage{
intervals = make_intervals(PF, PF_tolerance, eta, PT, PT_tolerance, zeta)
pf_matrix = intervals$pf_matrix
pt_matrix = intervals$pt_matrix
selected = select_highestprob_interval(dose_info, pf_matrix, pt_matrix)
best_pf = selected$best_pf
best_pt = selected$best_pt
decision_table[f_count+1, t_count+1] = make_decision(best_pf, best_pt, pf_matrix, pt_matrix,
                                                     PF = 0.2, PF_tolerance = 0.05, PT = 0.2, PT_tolerance = 0.05)
}
\arguments{
\item{best_pf}{Futility interval index for chosen model}

\item{best_pt}{Toxicity interval index for chosen model}

\item{pf_matrix}{Matrix containing start and end points of futility (pf) intervals}

\item{pt_matrix}{Matrix containing start and end points of toxicity (pt) intervals}

\item{PF}{Target tolerable futility probability}

\item{PF_tolerance}{Half-length of futility equivalence interval (epsilon_F)}

\item{PT}{Target tolerable toxicity probability}

\item{PT_tolerance}{Half-length of toxicity equivalence interval (epsilon_T)}
}
\value{
String containing decision for next dose ("D", "S", "E")
}
\description{
Make a decision to escalate (E), stay (S), or de-escalate (D) dose level based on the selected model.
Prioritizes safety over efficacy in that if the chosen model's estimated chance of toxicity is higher than the toxicity equivalence interval, choose to de-escalate.
Otherwise, choose to escalate if chosen model lies above the futility equivalence interval, and stay if it is within or below both equivalence intervals.
(Assumes that DU/EU decision, if it is the one to make, has already been taken care of before hitting this function.)
}
