% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_decision_table.R
\name{make_decision_table}
\alias{make_decision_table}
\title{Make Decision Table}
\usage{
make_decision_table(15, PF = 0.2, PF_tolerance = 0.05, eta = 0.8, PT = 0.2, PT_tolerance = 0.05, zeta = 0.8)
}
\arguments{
\item{N}{Number of patients tested at a dose}

\item{PF}{Target tolerable futility probability}

\item{PF_tolerance}{Half-length of futility equivalence interval (epsilon_F)}

\item{eta}{Upper futility threshold beyond which invokes the futility rule}

\item{PT}{Target tolerable toxicity probability}

\item{PT_tolerance}{Half-length of toxicity equivalence interval (epsilon_T)}

\item{zeta}{Upper toxicity threshold beyond which invokes the safety rule}
}
\value{
A matrix containing the decision table.
Futility counts are on the row axis, and toxicity counts are on the column axis.
}
\description{
Produces the pre-smoothed decision table for a certain dose sample size using given parameters and thresholds.
The decision table should be generalizable to any dose level. Safety is valued over efficacy, so any estimates exceeding the
toxicity tolerance will call for de-escalation.

\strong{IMPORTANT:} The decision table must be named in the following format to run simulations: "decision_table#" where #=N.
}
