% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_intervals.R
\name{make_intervals}
\alias{make_intervals}
\title{Probability Intervals (Regions) Setup}
\usage{
intervals = make_intervals(PF=0.2, PF_tolerance=0.05, eta=0.8, PT=0.2, PT_tolerance=0.05, zeta=0.8)

pf_matrix = intervals$pf_matrix

pt_matrix = intervals$pt_matrix
}
\arguments{
\item{PF}{Target tolerable futility probability}

\item{PF_tolerance}{Half-length of futility equivalence interval (epsilon_F)}

\item{eta}{Upper futility threshold beyond which invokes the futility rule}

\item{PT}{Target tolerable toxicity probability}

\item{PT_tolerance}{Half-length of toxicity equivalence interval (epsilon_T)}

\item{zeta}{Upper toxicity threshold beyond which invokes the safety rule}
}
\value{
A list containing:
\itemize{
\item pf_matrix -- matrix containing start and end points of p_f intervals
\item pt_matrix -- Matrix containing start and end points of p_t intervals
}
}
\description{
Make probability intervals given the target futility and toxicity level threshold parameters
}
