% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploss.R
\name{ploss}
\alias{ploss}
\title{Calculate Loss (1-Utility)}
\usage{
ploss(value = 0.18, target = 0.2, tolerance = 0.05) #calculate loss for 0.18 when equivalence interval is [0.15,0.25]
}
\arguments{
\item{value}{Estimated probability of futility/toxicity}

\item{target}{Target chance of futility/toxicity}

\item{tolerance}{Half-length of the equivalence interval}
}
\value{
Calculated numeric loss for given value
}
\description{
Calculates loss for a given value with target and tolerance levels. Can be used for toxicity or futility.
Loss = 0 if value below target - tolerance, 1 if above target + tolerance, and linear from 0 to 1 between.
}
