% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_final_dose.R
\name{select_final_dose}
\alias{select_final_dose}
\title{Final Dose Selection}
\usage{
select_final_dose(dose_info, stopping_rule = FALSE, filter = 0.5)
}
\arguments{
\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}

\item{sample_size}{Cumulative sample size at trial end}

\item{max_samp_size}{Maximum sample size for trial}

\item{stopping_rule}{Whether or not a stopping rule was invoked}

\item{filter}{Posterior sum threshold to determine dose availability for final selection (pf+pt <= filter)}
}
\value{
The final dose recommendation, or -99 if no dose was selected
}
\description{
Select the final dose out of available doses that has the highest utility.
}
\details{
No final dose is recommended if no doses are available, any stopping rules have been invoked, or no dose meets the selection criteria of pf+pt <= filter (e.g. filter = 0.5).
The latter criteria is to exclude doses with too high toxicity or futility, but have not invoked safety or futility rules.
Another alternative is to exclude doses with estimated chances of toxicity and futility above the equivalence intervals, but this has been found to be too strict a criteria and results in
no dose recommendation too often through simulation studies.
}
