% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_highestprob_interval.R
\name{select_highestprob_interval}
\alias{select_highestprob_interval}
\title{Select Model}
\usage{
selected = select_highestprob_interval(dose_info, pf_matrix, pt_matrix)
best_pf = selected$best_pf
best_pt = selected$best_pt
}
\arguments{
\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}

\item{pf_matrix}{Matrix containing start and end points of futility (pf) intervals}

\item{pt_matrix}{Matrix containing start and end points of toxicity (pt) intervals}
}
\value{
A list containing:
\itemize{
\item post_probs -- a matrix of posterior probailities per region (model)
\item best_pf -- futility interval index for chosen model
\item best_pt -- toxicity interval index for chosen model
}
}
\description{
Select the probability region (model) with highest posterior probability, given current results. Note that for generalizability, this function currently calculates probabilities for regions completely falling within the pf+pt <=1 region.
}
