% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Dose-Response Table Setup}
\usage{
doses = c(60,80,100,120,140)
dose_info = setup(doses)
}
\arguments{
\item{doses}{Vector of ordered dose levels}
}
\value{
An empty data frame in format

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]:
\itemize{
\item dose_lvl -- dose level
\item availability -- whether dose is available for further testing
\item N_d -- sample size
\item X_d -- futility counts
\item Y_d -- efficacy counts
\item Z_d -- toxicity counts
\item pf -- estimated chance of futility
\item pe -- estimated chance of efficacy
\item pt -- estimated chance of toxicity
}
}
\description{
Set up dose response-tracking table documenting the current dose's availability status,
sample size, and futility, efficacy, and toxicity counts and corresponding probabilities.
This table should be updated every time new cohort responses are available.
}
