% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinom_density.R
\name{trinom_density}
\alias{trinom_density}
\title{Trinomial Density}
\usage{
trinom_density(0.4, 0.1, 4, 1, 10)
}
\arguments{
\item{x}{Futility probability (pf)}

\item{y}{Toxicity probability (pt)}

\item{f_count}{Futility count (x)}

\item{t_count}{Toxicity count (z)}

\item{N}{Sample size}
}
\value{
The Trinomial(N, Pf, Pt) density at (f_count, t_count)
}
\description{
Calculate Trinomial(N, Pf, Pt) density with given parameters
}
