% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table.R
\name{update_table}
\alias{update_table}
\title{Update Dose-Response Table}
\usage{
dose_info = change_dose_info(80, 1, 2, 0) #update dose 80 with an *additional* 1 futility and 2 efficacy counts
}
\arguments{
\item{dose}{A numeric dose level (assumed to be available)}

\item{dose_info}{Dataframe containing

[dose level | availability | N_d | X_d | Y_d | Z_d | pf | pe | pt]}

\item{x_d}{Number of subjects experiencing futility}

\item{y_d}{Number of subjects experiencing efficacy}

\item{z_d}{Number of subjects experiencing toxicity}
}
\value{
A dataframe containing updated dose-response table (\code{dose_info})
}
\description{
Update table with dose-response information from a new cohort. For the given dose level, add the new futility, efficacy, and toxicity counts to existing numbers.
}
